pyproject.toml
[tool.poetry]
name = "stashapp-tag-importer"
version = "0.1.0"
description = ""
authors = ["soundchaser128 <soundchaser128@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.10"
stashapp-tools = "^0.2.19"
black = "^22.12.0"
python-dotenv = "^0.21.0"
httpx = "^0.23.3"
ipython = "^8.11.0"
questionary = "^1.10.0"

[tool.poetry.dev-dependencies]
pytest = "^5.2"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
